requirements.txt
django-rest-framework==0.1.0
djoser==2.3.3
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
pillow==11.3.0
pre-commit==3.8.0
psycopg2-binary==2.9.10
pytest-django==4.11.1
python-dotenv==1.1.1
